requirements.txt
fastapi==0.87.0
starlette==0.21.0
pydantic==1.8.2
uvicorn==0.22.0
databases[sqlite]
pydantic[email]
bcrypt==4.0.1
black
pyjwt
email-validator==1.3.1
fastapi-mail==1.2.2
itsdangerous